TK-04 - Made some refactoring of component structure and code
