+++ src/app/account/account.component.html
@@ -43,7 +43,11 @@
       <div class="compliance">
         <h1>
           <span>Compliance & Documentation</span>
-          <a class="link" href="#">See history <i class="pi pi-icon pi-arrow-right" [style]="{'font-size': '0.75rem'}"></i></a>
+          <a class="link" href="#">
+            <span>See history</span>
+            <i class="pi pi-icon pi-arrow-right"
+               [style.font-size.rem]="0.75"></i>
+          </a>
         </h1>
         <app-compliance [fields]="accountData.compliance"></app-compliance>
       </div>
