+++ src/app/account/account.component.scss
@@ -25,6 +25,9 @@
       }
 
       .link {
+        display: flex;
+        align-items: baseline;
+        gap: 0.5rem;
         font-size: 1rem;
         color: $app-main-blue-color;
         text-decoration: none;
