+++ src/app/account/communication/communication.component.html
@@ -1,4 +1,4 @@
-<p-card class="app-communication" [style]="{'width': '100%', 'padding-bottom': '1.5rem'}">
+<p-card class="app-communication" [style]="{ width: '100%', paddingBottom: '1.5rem' }">
   <div class="header">
     <span class="input-filter">
       <input
