+++ src/app/account/communication/components/message-card/message-card.component.html
@@ -19,7 +19,7 @@
   <div class="actions">
     @if (message.attachments?.length) {
       <button class="btn btn-attach" [pTooltip]="attachmentName" tooltipPosition="right">
-        <i class="pi pi-paperclip" [style]="{'font-size': '0.85rem'}"></i>  {{ message.attachments?.length }} attachments
+        <i class="pi pi-paperclip" [style]="{ fontSize: '0.85rem' }"></i>  {{ message.attachments?.length }} attachments
       </button>
     }
     @if (message.canReply) {
