+++ src/app/account/communication/components/message-card/message-card.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles.scss";
+@import "../../../../../styles";
 
 .message-card {
   display: flex;
