+++ src/app/account/performance-metrics/components/performance-card/performance-card.component.html
@@ -1,6 +1,6 @@
 <p-card>
   <div class="app-performance-card"
-       [style]="{'min-height': performanceData.bars?.length ? 'auto' : '220px'}">
+       [style]="{ minHeight: performanceData.bars?.length ? 'auto' : '220px' }">
     <div>
       <div class="card-title">
         <h3>{{ performanceData.title }}</h3>
@@ -30,7 +30,7 @@
       </div>
     } @else {
       <a class="link" [href]="performanceData.link">
-        {{ performanceData.linkTitle }} <i class="pi pi-icon pi-arrow-right" [style]="{'font-size': '0.75rem'}"></i>
+        {{ performanceData.linkTitle }} <i class="pi pi-icon pi-arrow-right" [style]="{ fontSize: '0.75rem' }"></i>
       </a>
     }
   </div>
