+++ src/app/account/performance-metrics/components/performance-card/performance-card.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles.scss";
+@import "../../../../../styles";
 
 .app-performance-card {
   display: flex;
