+++ src/app/account/performance-metrics/performance-metrics.component.ts
@@ -1,13 +1,11 @@
 import { Component, Input } from '@angular/core';
-import { PerformanceCardComponent } from '../../shared/components/performance-card/performance-card.component';
-import { NgStyle } from '@angular/common';
+import { PerformanceCardComponent } from './components/performance-card/performance-card.component';
 import { PerformanceCard } from '../../shared/types/account.types';
 
 @Component({
   selector: 'app-performance-metrics',
   imports: [
-    PerformanceCardComponent,
-    NgStyle
+    PerformanceCardComponent
   ],
   templateUrl: './performance-metrics.component.html',
   styleUrl: './performance-metrics.component.scss',
