+++ src/app/account/policies-cards/components/policy-card/policy-card.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles.scss";
+@import "../../../../../styles";
 
 .app-policy-card {
   padding: 1.75rem;
