+++ src/app/account/policies-grid/policies-grid.component.html
@@ -1,4 +1,4 @@
-<p-card class="app-policies-grid" [style]="{'width': '100%'}">
+<p-card class="app-policies-grid" [style]="{ width: '100%' }">
   <div class="header">
     <span class="input-filter">
       <input type="text"
