+++ src/app/dashboard/market-intelligence/market-intelligence.component.html
@@ -1,6 +1,6 @@
 <p-card class="app-market-card"
-        [style]="{'width': 'var(--top-right-column-width)',
-                  'height': 'calc(var(--top-row-height) - var(--top-right-column-height) - var(--layout-gap)'}">
+        [style]="{ width: 'var(--top-right-column-width)',
+                   height: 'calc(var(--top-row-height) - var(--top-right-column-height) - var(--layout-gap)' }">
   <h2>Market intelligence</h2>
   <div class="content">
     <ul class="market-list">
