+++ src/app/dashboard/my-accounts/my-accounts.component.html
@@ -1,4 +1,4 @@
-<p-card class="app-my-accounts" [style]="{'width': '100%'}">
+<p-card class="app-my-accounts" [style]="{ width: '100%' }">
   <div class="header">
     <h2>My accounts</h2>
