+++ src/app/dashboard/portfolio-goals/portfolio-goals.component.html
@@ -1,4 +1,5 @@
-<p-card class="app-portfolio-goals" [style]="{'width': 'var(--top-right-column-width)', 'height': 'var(--top-row-height)'}">
+<p-card class="app-portfolio-goals" [style]="{ width: 'var(--top-right-column-width)',
+                                               height: 'var(--top-row-height)' }">
   <h2>Portfolio goals</h2>
   <div class="content">
     @if (isLoading) {
