+++ src/app/dashboard/quick-actions/quick-actions.component.html
@@ -1,4 +1,4 @@
-<p-card class="app-quick-actions" [style]="{'width': 'var(--top-right-column-width)', 'height': 'var(--top-right-column-height)'}">
+<p-card class="app-quick-actions" [style]="{ width: 'var(--top-right-column-width)', height: 'var(--top-right-column-height)'}">
   <h2>Quick actions</h2>
   <div class="content">
     @for (action of quickActions; track action) {
