+++ src/app/dashboard/work-queue/work-queue.component.html
@@ -1,4 +1,4 @@
-<p-card class="app-work-queue" [style]="{'width': '100%', 'height': 'var(--top-row-height)'}" styleClass="app-work-queue">
+<p-card class="app-work-queue" [style]="{ width: '100%', height: 'var(--top-row-height)' }" styleClass="app-work-queue">
   <div class="header">
     <h2>Work Queue</h2>
   </div>
