+++ src/app/header/breadcrumbs/breadcrumbs.component.html
@@ -1,7 +1,7 @@
 <div class="app-breadcrumbs">
   <p-breadcrumb [model]="items"
                 [home]="home"
-                [style]="{'padding': '0', 'background-color': 'transparent'}">
+                [style]="{ padding: '0', backgroundColor: 'transparent' }">
     <ng-template #item let-item>
       <a class="p-breadcrumb-item-link"
          [ngClass]="{'active': item.active}"
