+++ src/app/shared/components/goal-bar/goal-bar.component.ts
@@ -1,12 +1,9 @@
 import { Component, Input } from '@angular/core';
-import { NgStyle } from '@angular/common';
 import { GoalBarData } from '../../types/bar.types';
 
 @Component({
   selector: 'app-goal-bar',
-  imports: [
-    NgStyle
-  ],
+  imports: [],
   templateUrl: './goal-bar.component.html',
   styleUrl: './goal-bar.component.scss',
   standalone: true
