+++ src/app/shared/components/target-bar/target-bar.component.ts
@@ -1,12 +1,9 @@
 import { Component, Input } from '@angular/core';
 import { TargetBarData } from '../../types/bar.types';
-import { NgStyle } from '@angular/common';
 
 @Component({
   selector: 'app-target-bar',
-  imports: [
-    NgStyle
-  ],
+  imports: [],
   templateUrl: './target-bar.component.html',
   styleUrl: './target-bar.component.scss',
   standalone: true
